backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Database
motor==3.4.0                  # async MongoDB driver
pymongo==4.7.2

# AI / LangGraph / RAG
langchain==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
langgraph==0.1.1
openai==1.30.1
chromadb==0.5.0
tiktoken==0.7.0
google-generativeai==0.5.4

# Market Data
yfinance==0.2.38
pandas==2.2.2

# News / Sentiment
httpx==0.27.0
newsapi-python==0.2.7

# Utilities
python-dateutil==2.9.0
pytz==2024.1
